// dashboard/ui/src/api/generated/models/analysisJobRequestAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */

export type AnalysisJobRequestAnalysis = typeof AnalysisJobRequestAnalysis[keyof typeof AnalysisJobRequestAnalysis];


export const AnalysisJobRequestAnalysis = {
  apl: 'apl',
  clusters: 'clusters',
  density: 'density',
  dist: 'dist',
  energies: 'energies',
  hbonds: 'hbonds',
  inter: 'inter',
  linter: 'linter',
  lorder: 'lorder',
  pairwise: 'pairwise',
  pca: 'pca',
  perres: 'perres',
  pockets: 'pockets',
  rgyr: 'rgyr',
  rmsds: 'rmsds',
  rmsf: 'rmsf',
  sas: 'sas',
  thickness: 'thickness',
  tmscore: 'tmscore',
} as const;